# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: "Описание спецификации всех API методов, используемых в online diffusion\
    \ сервисе"
  title: Online diffusion API
  version: 0.0.1
  x-logo:
    url: https://i.imgur.com/1kUh580.jpg
...
servers:
- description: "Локальный сервер, актуально для запуска на своем компьютере"
  url: http://localhost:8080
  x-last-modified: 1698183338768
security:
- JWTAuth: []
...
paths:
...
  /api/v1/profile:
    get:
      description: ""
      operationId: get-profile
      parameters: []
      responses:
        "200":
          content: {}
          description: "все хорошо, информация по юзеру получена"
      security:
      - JWTAuth: []
      summary: Получаем инфо из профиля
      tags:
      - User
      x-accepts: application/json
      x-tags:
      - tag: User
  /api/v1/profile/username:
    put:
      description: ""
      operationId: profile-username
      parameters:
      - description: Имя пользователя
        explode: true
        in: query
        name: username
        required: true
        schema:
          example: user
          maxLength: 32
          minLength: 4
          type: string
        style: form
      responses:
        "200":
          content: {}
          description: "все хорошо, имя пользователя было успешно изменено"
        "400":
          description: "изменение имени пользователя на данное имя невозможно, так\
            \ как не удовлетворяет требованиям к уникальности"
      security:
      - JWTAuth: []
      summary: Меняем имя юзеру
      tags:
      - User
      x-accepts: application/json
      x-tags:
      - tag: User
  /api/v1/profile/password:
    put:
      description: ""
      operationId: profile-password
      parameters:
      - description: Пароль пользователя
        explode: true
        in: query
        name: password
        required: true
        schema:
          example: querty123
          maxLength: 32
          minLength: 8
          type: string
        style: form
      responses:
        "200":
          content: {}
          description: "все хорошо, пароль был успешно изменен"
      security:
      - JWTAuth: []
      summary: Меняем пароль юзеру
      tags:
      - User
      x-accepts: application/json
      x-tags:
      - tag: User
  /api/v1/profile/gallery:
    get:
      description: ""
      operationId: profile-gallery
      parameters:
      - description: "Маркер, показывающий начиная с какого id подгружать ресурсы"
        explode: true
        in: query
        name: marker
        required: false
        schema:
          example: 1
          format: int64
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryPagingWrapper'
          description: "все хорошо, присланы изображения галереи"
        "404":
          description: такой маркер не найден
      security:
      - JWTAuth: []
      summary: Получить галерею
      tags:
      - User
      x-accepts: application/json
      x-tags:
      - tag: User
      - tag: Images
components:
...
  schemas:
...
    GalleryPagingWrapper:
      description: Оболочка для пагинации
      example:
        nextMarker: 0.8008281904610115
        images:
        - seed: seed
          authorName: authorName
          authorAvatarUrl: authorAvatarUrl
          photoId: 0.8008281904610115
          anti-prompt: anti-prompt
          model: model
          prompt: prompt
        - seed: seed
          authorName: authorName
          authorAvatarUrl: authorAvatarUrl
          photoId: 0.8008281904610115
          anti-prompt: anti-prompt
          model: model
          prompt: prompt
      properties:
        nextMarker:
          description: Маркер для следующего запроса
          type: number
        images:
          description: Список присланных изображений в нужном порядке
          items:
            $ref: '#/components/schemas/Image'
          type: array
      type: object
...
    Image:
      description: Объект с информацией об фотографии
      example:
        seed: seed
        authorName: authorName
        authorAvatarUrl: authorAvatarUrl
        photoId: 0.8008281904610115
        anti-prompt: anti-prompt
        model: model
        prompt: prompt
      properties:
        photoId:
          description: Id фотографии
          type: number
        authorName:
          description: Имя автора
          type: string
        authorAvatarUrl:
          description: URL аватара автора
          type: string
        prompt:
          description: prompt использовшийся для создания изображения
          type: string
        anti-prompt:
          description: anti-prompt использовшийся для создания изображения
          type: string
        seed:
          description: "seed, использовшийся для создания изображения"
          type: string
        model:
          description: "Модель, использувшееся для создания изображения"
          type: string
      required:
      - anti-prompt
      - authorAvatarUrl
      - authorName
      - model
      - photoId
      - prompt
      - seed
      type: object
...
  securitySchemes:
    JWTAuth:
      bearerFormat: JWT
      description: "Example: \n> Authorization: Bearer <token>"
      scheme: bearer
      type: http
      x-last-modified: 1698183246392